LayoutTests/http/tests/security/XFrameOptions/cross-origin-iframe-post-form-to-parent-same-origin-x-frame-options-page-allow-expected.txt

There should be content in the iframe below



--------
Frame: '<!--frame1-->'
--------
PASS: This should show up as the parent is in the same origin.
